# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "Registrar cuenta (rol USUARIO)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cuenta creada y sesión iniciada"
          }
        },
        "tags": [
          "Autenticación"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Iniciar sesión: devuelve un JWT (Bearer) con claim scope",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Autenticación"
        ]
      }
    },
    "/api/v1/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgot",
        "summary": "Solicitar enlace de recuperación (respuesta siempre genérica)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Autenticación"
        ]
      }
    },
    "/api/v1/auth/reset-password": {
      "post": {
        "operationId": "AuthController_reset",
        "summary": "Restablecer la clave con el token recibido",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Autenticación"
        ]
      }
    },
...
    "/api/v1/me": {
      "get": {
        "operationId": "CuentaController_me",
        "summary": "Mi perfil",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Mi cuenta"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "CuentaController_actualizar",
        "summary": "Actualizar mi perfil (datos de facturación incluidos)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarPerfilDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Mi cuenta"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/me/orders": {
      "get": {
        "operationId": "CuentaController_misOrdenes",
        "summary": "Mis reservas",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Mi cuenta"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/orders/{id}/invoice": {
      "get": {
        "operationId": "CuentaController_factura",
        "summary": "Factura simulada de la reserva (v_factura)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Mi cuenta"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/orders/{id}/timeline": {
      "get": {
        "operationId": "CuentaController_timeline",
        "summary": "Trazabilidad de la reserva (dueño, anfitrión o ADMIN)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Mi cuenta"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/orders/{id}/cancel-preview": {
      "get": {
        "operationId": "CuentaController_previewCancelacion",
        "summary": "Previsualizar penalidad y reembolso de una cancelación",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Mi cuenta"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/orders/{id}/review": {
      "post": {
        "operationId": "CuentaController_resena",
        "summary": "Reseñar una estancia completada",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearResenaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Mi cuenta"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Posada EC · Alojamientos API",
    "description": "Implementación del contrato \"GDS Alojamientos Core API\" (rutas del YAML bajo /api/v1) y rutas propias (auth, cuenta, anfitrión, admin). El ownerId de cada reserva se infiere del `sub` del JWT. OAuth2 simplificado: el token se obtiene con POST /api/v1/auth/login y se envía como Bearer.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Este servidor"
    }
  ],
  "components": {
    "securitySchemes": {
      "OAuth2Security": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://auth.booking-hub.com/oauth2/authorize",
            "tokenUrl": "https://auth.booking-hub.com/oauth2/token",
            "scopes": {
              "alojamientos:read": "Leer información de alojamientos, disponibilidades y reservas",
              "alojamientos:book": "Crear y alterar reservas",
              "alojamientos:cancel": "Cancelar reservas",
              "alojamientos:webhooks": "Gestionar webhooks"
            }
          },
          "clientCredentials": {
            "tokenUrl": "https://auth.booking-hub.com/oauth2/token",
            "scopes": {
              "alojamientos:read": "Leer información de alojamientos, disponibilidades y reservas",
              "alojamientos:book": "Crear y alterar reservas",
              "alojamientos:cancel": "Cancelar reservas",
              "alojamientos:webhooks": "Gestionar webhooks"
            }
          }
        },
        "description": "OAuth2 del contrato. En este prototipo el token Bearer se emite en /api/v1/auth/login."
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 160
          },
          "password": {
            "type": "string",
            "pattern": "POLITICA_CLAVE"
          },
          "nombres": {
            "type": "string",
            "minLength": 2,
            "maxLength": 80,
            "pattern": "NOMBRE_PERSONA"
          },
          "apellidos": {
            "type": "string",
            "minLength": 2,
            "maxLength": 80,
            "pattern": "NOMBRE_PERSONA"
          },
          "telefono": {
            "type": "string",
            "pattern": "TELEFONO_EC"
          },
          "tipo_documento": {
            "type": "string",
            "enum": [
              "CEDULA",
              "RUC",
              "PASAPORTE"
            ]
          },
          "numero_documento": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "nombres",
          "apellidos"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "maxLength": 72
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Token de 32 bytes en hexadecimal recibido por \"correo\"",
            "minLength": 64,
            "maxLength": 64
          },
          "password": {
            "type": "string",
            "pattern": "POLITICA_CLAVE"
          }
        },
        "required": [
          "token",
          "password"
        ]
      },
...
      "ActualizarPerfilDto": {
        "type": "object",
        "properties": {
          "nombres": {
            "type": "string",
            "minLength": 2,
            "maxLength": 80,
            "pattern": "NOMBRE_PERSONA"
          },
          "apellidos": {
            "type": "string",
            "minLength": 2,
            "maxLength": 80,
            "pattern": "NOMBRE_PERSONA"
          },
          "telefono": {
            "type": "string",
            "pattern": "TELEFONO_EC"
          },
          "tipo_documento": {
            "type": "string",
            "enum": [
              "CEDULA",
              "RUC",
              "PASAPORTE"
            ]
          },
          "numero_documento": {
            "type": "string"
          },
          "razon_social": {
            "type": "string",
            "maxLength": 160
          }
        }
      },
      "CrearResenaDto": {
        "type": "object",
        "properties": {
          "nota_global": {
            "type": "number",
            "description": "Nota global de 1 a 10",
            "minimum": 1,
            "maximum": 10
          },
          "comentario": {
            "type": "string",
            "description": "Comentario opcional (10 a 1000 caracteres). Se valida automáticamente en el servicio:\nsin un mismo carácter 4+ veces seguidas, sin enlaces, correos ni teléfonos y sin groserías.",
            "minLength": 10,
            "maxLength": 1000
          }
        },
        "required": [
          "nota_global"
        ]
      },
...
    }
  }
}